+++ Pluminus/View/MapView.swift
@@ -1,87 +1,68 @@
 //
-//  MapView.swift
+//  MapFocusView.swift
 //  Pluminus
 //
 //  Created by kimsangwoo on 2023/08/22.
-// test
+//
 
 import SwiftUI
- 
-struct MapView: View {
+import MapKit
+
+struct MapView: UIViewRepresentable {
     @Binding var countryName: String
-    @Binding var continent: String
     @Binding var locality: String
     
-    @Environment(\.dismiss) var dismiss
+    func makeCoordinator() -> MapViewDelegate {
+        return MapViewDelegate()
+    }
     
-    var body: some View {
-        ZStack(alignment: .top) {
-            MapFocusView(
-                countryName: $countryName,
-                locality: $locality
-            )
-            
-            VStack {
-                HStack {
-                    Spacer()
-                    
-                    Button(action: {
-                        HapticManager.instance.impact(style: .light)
-                        dismiss()
-                    }, label: {
-                        ZStack {
-                            Rectangle()
-                                .frame(width: 60, height: 60)
-                                .foregroundStyle(.clear)
-                            Circle()
-                                .frame(width: 28, height: 28)
-                                .foregroundStyle(.clear)
-                                .background(.regularMaterial)
-                                .clipShape(Circle())
-                                .overlay(
-                                    Circle()
-                                        .stroke(.regularMaterial, lineWidth: 1)
-                            )
-                            Image(systemName: "xmark")
-                                .resizable()
-                                .bold()
-                                .foregroundStyle(.orange)
-                                .frame(width: 12, height: 12)
-                        }
-                    })
-                }
+    func makeUIView(context: Context) -> MKMapView {
+        let mapView = MKMapView()
+        mapView.delegate = context.coordinator
+        return mapView
+    }
+    
+    func updateUIView(_ uiView: MKMapView, context: Context) {
+        let geocoder = CLGeocoder()
+        let country = locality.isEmpty ? countryName : locality
+        
+        geocoder.geocodeAddressString(country) { placemarks, error in
+            if let placemark = placemarks?.first,
+               let location = placemark.location {
+                let coordinate = location.coordinate
+                let region = MKCoordinateRegion(center: coordinate, span: MKCoordinateSpan(latitudeDelta: 10, longitudeDelta: 10))
+                uiView.setRegion(region, animated: false)
                 
-                Spacer()
-                
-                VStack(alignment: .leading) {
-                    VStack(alignment: .leading, spacing: 0) {
-                        Text("\(countryName)")
-                            .font(.system(size: 24, weight: .bold))
-                            .foregroundColor(.primary)
-                        
-                        Text("\(locality)")
-                            .font(.system(size: 24, weight: .bold))
-                            .foregroundColor(.secondary)
-                            .padding(.bottom, 8)
-                        
-                        Text("\(continent)")
-                            .font(.system(size: 14, weight: .semibold))
-                            .foregroundColor(.secondary)
-                    }
-                    .padding(.top, 20)
-                    .padding(.horizontal, 24)
-                }
-                .frame(width: screenWidth, alignment: .leading)
-                .padding(.bottom, 60)
-                .background(.regularMaterial)
-                .clipShape(RoundedRectangle(cornerRadius: 16))
-                .overlay(
-                    RoundedRectangle(cornerRadius: 16)
-                        .stroke(.regularMaterial, lineWidth: 1)
-                )
-                .shadow(color: .black.opacity(0.2), radius: 20)
-            } //VStack
-        } // ZStack
-        .ignoresSafeArea()
-    } // body
-} // struct
+                let annotation = MKPointAnnotation()
+                annotation.coordinate = coordinate
+                annotation.title = country
+                uiView.addAnnotation(annotation)
+            }
+        }
+    }
+}
+
+class MapViewDelegate: NSObject, MKMapViewDelegate {
+    func mapView(_ mapView: MKMapView, viewFor annotation: MKAnnotation) -> MKAnnotationView? {
+        guard annotation is MKPointAnnotation else { return nil } // Check if it's an MKPointAnnotation
+
+        let identifier = "CustomAnnotationView"
+
+        var annotationView = mapView.dequeueReusableAnnotationView(withIdentifier: identifier)
+        if annotationView == nil {
+            annotationView = MKMarkerAnnotationView(annotation: annotation, reuseIdentifier: identifier)
+            annotationView?.canShowCallout = true
+
+            if let markerAnnotationView = annotationView as? MKMarkerAnnotationView {
+                markerAnnotationView.markerTintColor = .red
+                markerAnnotationView.glyphText = "●"
+                markerAnnotationView.glyphTintColor = .white
+            }
+        } else {
+            annotationView?.annotation = annotation
+        }
+
+        return annotationView
+    }
+}
+
